# service/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenID Federation Registry API",
    "description": "REST API for managing OpenID Federation entities, policies, modules, and trust marks. This API provides endpoints for creating, reading, updating, and deleting federation entities, policies, trust anchors, resolvers, trust marks, and trust mark subjects.",
    "contact": {
      "name": "Sweden Connect",
      "url": "https://www.swedenconnect.se",
      "email": "support@swedenconnect.se"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.5.10-SNAPSHOT"
  },
  "servers": [
    {
      "url": "https://registry.local.swedenconnect.se",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/registry/v1/{tenant}/{orgNumber}/subordinates/{subordinateId}": {
      "get": {
        "tags": [
          "Subordinates"
        ],
        "summary": "Get subordinate by id",
        "operationId": "getSubordinate",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "subordinateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subordinate"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Subordinates"
        ],
        "summary": "Update subordinate",
        "operationId": "updateSubordinate",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "subordinateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subordinate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subordinate"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Subordinates"
        ],
        "summary": "Create subordinate with specified ID",
        "operationId": "createSubordinateWithId",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "subordinateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subordinate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subordinate"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Subordinates"
        ],
        "summary": "Delete subordinate",
        "operationId": "deleteSubordinate",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          },
          {
            "name": "subordinateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/registry/v1/{tenant}/{orgNumber}/subordinates/": {
      "post": {
        "tags": [
          "Subordinates"
        ],
        "summary": "Create subordinate with auto-generated ID",
        "operationId": "createSubordinate",
        "parameters": [
          {
            "name": "tenant",
            "in": "path",
            "description": "Tenant identifier, also the function name",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "swedenconnect"
          },
          {
            "name": "orgNumber",
            "in": "path",
            "description": "Organization number",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2021003948"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subordinate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subordinate"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Subordinate": {
        "type": "object",
        "properties": {
          "subordinateId": {
            "type": "string",
            "format": "uuid",
            "description": "Subordinate ID",
            "readOnly": true
          },
          "taImId": {
            "type": "string",
            "format": "uuid",
            "description": "TaIm ID that this subordinate belongs to"
          },
          "jwks": {
            "type": "object",
            "additionalProperties": {},
            "description": "JWKSet for Subordinate",
            "example": "JSON Web Key Set"
          },
          "entityIdentifier": {
            "type": "string",
            "description": "Subject (entity identifier)",
            "example": "https://subordinate.example.se"
          },
          "crit": {
            "type": "array",
            "description": "List of crit claims",
            "example": [
              "claim1",
              "claim2"
            ],
            "items": {
              "type": "string"
            }
          },
          "metadataPolicyCrit": {
            "type": "array",
            "description": "List of metadata_policy_crit",
            "example": [
              "operator1",
              "operator2"
            ],
            "items": {
              "type": "string"
            }
          },
          "ecLocation": {
            "type": "string",
            "description": "ecLocation - location where the actual entity statement is placed. Expressed as a url. Ex https://my.company.se/entitystatement"
          },
          "ecLocationAutomaticResolve": {
            "type": "boolean",
            "description": "When true, eclocation will be loaded from the hosted entity with the same issuer entityid"
          },
          "effectiveEcLocation": {
            "type": "string",
            "description": "Effective EcLocation that is calculated serverside",
            "readOnly": true
          },
          "metadataPolicy": {
            "type": "object",
            "additionalProperties": {},
            "description": "Metadata policy for this subordinate statement. Optional. Must be valid JSON if provided.",
            "example": {
              "openid_provider": {
                "subject_types_supported": {
                  "value": [
                    "pairwise"
                  ]
                }
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT token obtained from authentication endpoint",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}